+++ client/src/locales/strings.json
@@ -135,6 +135,10 @@
         "bioTextArea": {
           "label": "Biografía",
           "placeholder": "Escribe tu Biografía"
+        },
+        "socialNetworks": {
+          "label": "Redes sociales",
+          "validateText": "Escribe un enlace correcto"
         }
       }
     }
